--- platform/linux-generic/pktio/dpdk_parse.c
@@ -1,91 +1,0 @@
-/* Copyright (c) 2018, Linaro Limited
- * Copyright (c) 2021-2022, Nokia
- * All rights reserved.
- *
- * SPDX-License-Identifier:     BSD-3-Clause
- */
-
-#include <odp/autoheader_internal.h>
-
-#ifdef _ODP_PKTIO_DPDK
-
-#include <odp_posix_extensions.h>
-
-#include <odp_parse_internal.h>
-#include <odp_packet_io_internal.h>
-#include <odp_packet_dpdk.h>
-#include <odp/api/byteorder.h>
-#include <odp/api/plat/byteorder_inlines.h>
-
-#include <protocols/eth.h>
-#include <protocols/udp.h>
-#include <protocols/tcp.h>
-
-#include <rte_config.h>
-#include <rte_mbuf.h>
-#include <rte_version.h>
-
-#if RTE_VERSION < RTE_VERSION_NUM(21, 11, 0, 0)
-	#define RTE_MBUF_F_RX_IP_CKSUM_MASK PKT_RX_IP_CKSUM_MASK
-	#define RTE_MBUF_F_RX_IP_CKSUM_GOOD PKT_RX_IP_CKSUM_GOOD
-	#define RTE_MBUF_F_RX_IP_CKSUM_UNKNOWN PKT_RX_IP_CKSUM_UNKNOWN
-	#define RTE_MBUF_F_RX_L4_CKSUM_MASK PKT_RX_L4_CKSUM_MASK
-	#define RTE_MBUF_F_RX_L4_CKSUM_GOOD PKT_RX_L4_CKSUM_GOOD
-	#define RTE_MBUF_F_RX_L4_CKSUM_UNKNOWN PKT_RX_L4_CKSUM_UNKNOWN
-#endif
-
-#define IP4_CSUM_RESULT(ol_flags) ((ol_flags) & RTE_MBUF_F_RX_IP_CKSUM_MASK)
-#define L4_CSUM_RESULT(ol_flags) ((ol_flags) & RTE_MBUF_F_RX_L4_CKSUM_MASK)
-
-/**
- * DPDK packet parser
- */
-int _odp_dpdk_packet_parse_common(odp_packet_hdr_t *pkt_hdr, const uint8_t *ptr,
-				  uint32_t frame_len, uint32_t seg_len,
-				  struct rte_mbuf *mbuf, int layer,
-				  odp_pktin_config_opt_t pktin_cfg)
-{
-	packet_parser_t *prs = &pkt_hdr->p;
-	uint64_t mbuf_ol = mbuf->ol_flags;
-
-	if (odp_unlikely(layer == ODP_PROTO_LAYER_NONE))
-		return 0;
-
-	/* Assume valid L2 header, no CRC/FCS check in SW */
-	prs->l2_offset = 0;
-
-	if (layer >= ODP_PROTO_LAYER_L3) {
-		int ip_chksum = IP4_CSUM_RESULT(mbuf_ol);
-
-		if (ip_chksum == RTE_MBUF_F_RX_IP_CKSUM_GOOD) {
-			prs->input_flags.l3_chksum_done = 1;
-		} else if (ip_chksum != RTE_MBUF_F_RX_IP_CKSUM_UNKNOWN) {
-			prs->input_flags.l3_chksum_done = 1;
-			prs->flags.l3_chksum_err = 1;
-		}
-	}
-
-	if (layer >= ODP_PROTO_LAYER_L4) {
-		int l4_chksum = L4_CSUM_RESULT(mbuf_ol);
-
-		if (l4_chksum == RTE_MBUF_F_RX_L4_CKSUM_GOOD) {
-			prs->input_flags.l4_chksum_done = 1;
-		} else if (l4_chksum != RTE_MBUF_F_RX_L4_CKSUM_UNKNOWN) {
-			prs->input_flags.l4_chksum_done = 1;
-			prs->flags.l4_chksum_err = 1;
-		}
-	}
-
-	pktin_cfg.bit.ipv4_chksum = 0;
-	pktin_cfg.bit.udp_chksum = 0;
-	pktin_cfg.bit.tcp_chksum = 0;
-	pktin_cfg.bit.sctp_chksum = 0;
-
-	return _odp_packet_parse_common(pkt_hdr, ptr, frame_len, seg_len, layer,
-					pktin_cfg);
-}
-
-#else /* _ODP_PKTIO_DPDK */
-/* Avoid warning about empty translation unit */
-typedef int _odp_dummy;
-#endif
